+++ silx/gui/_glutils/Context.py
