+++ silx/gui/_glutils/FramebufferTexture.py
