+++ silx/gui/_glutils/Texture.py
