+++ silx/gui/_glutils/utils.py
