+++ silx/gui/plot3d/setup.py
@@ -32,7 +32,6 @@ from numpy.distutils.misc_util import Configuration
 
 def configuration(parent_package='', top_path=None):
     config = Configuration('plot3d', parent_package, top_path)
-    config.add_subpackage('glutils')
     config.add_subpackage('scene')
     config.add_subpackage('test')
     config.add_subpackage('utils')
