+++ silx/gui/plot3d/test/__init__.py
@@ -55,12 +55,10 @@ def suite():
 
     # Import here to avoid loading modules if tests are disabled
 
-    # from ..glutils import test as test_glutils
     from ..scene import test as test_scene
     from ..utils import test as test_utils
 
     test_suite = unittest.TestSuite()
-    # test_suite.addTest(test_glutils.suite())
     test_suite.addTest(test_scene.suite())
     test_suite.addTest(test_utils.suite())
     return test_suite
