+++ silx/gui/setup.py
@@ -32,6 +32,7 @@ from numpy.distutils.misc_util import Configuration
 
 def configuration(parent_package='', top_path=None):
     config = Configuration('gui', parent_package, top_path)
+    config.add_subpackage('_glutils')
     config.add_subpackage('qt')
     config.add_subpackage('plot')
     config.add_subpackage('fit')
